refactored atleast some of the code in meme and selfie to own module, photo-util, also api calls to api
